move protocols to viewModel
